Refactor for precise user value memory management
This diff adds a `.destructor` callback to the user value received by
ArbiterCreateProjectIdentifier(). This change is to provide precise lifetime
control of those user values.

This change isn't as simple as I would have liked it to be. The crux of these
changes it that the user value is now additionally stored in the Ruby data
object, so that the `.destructor` can nil it out. This one change resulted in
most of the changes in this commit:

1. Introduction of `ProjectIdentifierData`
2. User value `.data` changed from `VALUE` to `VALUE *`