Split graph addition method into node adding and edge adding
This helps surface and resolve some issues where we would clobber the
_dependents_ of a given node, because we baked in a lot of assumptions
that there would only be one (per code path).

Since there’s no reason it has to be part of the behavior of addNode(),
let’s just separate the two, and make it clear that edges can be added
arbitrarily without otherwise changing the node.